+++ CHANGES.rst
@@ -6,6 +6,8 @@ develop
 -------
 - add lst-sattracker application
 - fix WAMP serialization error when publishing binary data (e.g. "char 0x9c") by using MsgPack serialization
+- augment c source file before compilation re. ``#include "mbed.h"`` vs. ``#include "stdint.h"``
+
 
 .. _v0.4.0: